bug fixes to correct memory leaks in the VE-Open code as well as some of the corba code. The major problem with our code is that we do things like:
char* temp = <some function that returns a char*>
std::string var( temp );

which could also be:

std::string var( <some function that returns a char*> );

In both of these cases if the function is passing ownership of the memory to the caller the above code causes a memory leak because string copies the char*. This means we need to delete the char* that is passed. I have corrected the few that I have found but I am sure we have this problem in many places in our code. It will probably be especially common in the XML and CORBA code.