Fix build issues with CMake 2.8.11.2 on Mac OS X 10.8
The behavior of FindBoost.cmake seems to have changed in the most recent
CMake releases, such that each find_package( Boost ) call overwrites the
contents of the Boost_LIBRARIES variable, rather than appending to it.
The result is that on the Mac, this variable won't contain all the Boost
libraries needed to link libves_xplorer_scenegraph (the build complains
about unresolved symbols for the Boost regex library). My solution is to
manually specify the Boost dependencies instead of relying on the
Boost_LIBRARIES variable supplying the right information. GNU's linker
must be smarter than OS X's linker, since this CMake behavior change
doesn't seem to cause problems on Linux. 



Former-commit-id: c9f35db66ca750e8df095bba9c168c83e3d8a513