fix Visual Studio-isms in the code
PolydataPropertySet.h:

Member functions don't need to be qualified with the class name in the class definition.

PreferencesPropertySet.cxx:

Converting an rvalue to a non-const reference works in Visual Studio (and possibly also clang), but it isn't standard C++. Fix this by creating the propertystore::PropertyPtr objects on the stack first, then pass them in to the functions.