add workaround for GCC C++11 & Qt connect() macro conflict
The way old Qt releases define the connect() macro for signal & slot connections is perfectly valid in C++03, but causes problems with C++11.

See "User defined literals and whitespace" at http://gcc.gnu.org/gcc-4.7/porting_to.html

The workaround is to use the set_source_files_properties() CMake command to set the "-std=c++03" compiler flag for Qt-related source files, which will override the "-std=c++11" flag that normally needs to be set at the project level. This is made conditional on a check for the GNU compiler, since we have not yet seen any problems on Mac OS X or Windows. Presumably, this is because clang and MSVC either have workarounds already in place, or they are just more forgiving than GCC.