fix the Preferences tab layout
Correct the layout such that the "Save Camera State" QPushButton is a
child of the QVBoxLayout. Before, the button was a child of the
ves::conductor::VesTreePropertyBrowser widget. As a fluke, this
happened to produce the correct visual output with Qt 4.8 on Windows, but didn't
work right with Qt 4.6 on RHEL6.