fix head tracker <-> gamepad nav inconsistency
Add a fix for the issue observed at the Rock Island Arsenal where if you
rotated the world using the gamepad navigation, any subsequent head
motions would have weird results that were inconsistent with the navigation you had done.

The root of the issue is that head tracker data is in a fixed coordinate
frame (relative to the CAVE), but rotations coming from gamepad nav can
change that coordinate frame. This commit changes how the nav results
and head tracker data are combined to update the view: the orientation
information from the gamepad nav is extracted, and the inverse
orientation is applied to the head tracker positions to keep things
consistent.