more cleanup of kludgy preference loading code
Rip out a bunch of preference handling code that was scattered across a few (unrelated) classes.

In its place, put in a slot that subscribes to the "VesFileLoaded" signal that is generated by the ves::conductor::NetworkLoader class when it's done loading a .ves file. In the slot, use the DatabaseManager to check if there are any XplorerPreferences associated with the file. If so, use those stored preferences to populate a new PreferencesPropertySet object, and hand it off to the preferences UI.

This approach works, but could probably stand to be refactored. Currently, the slot is implemented in the PreferencesTab UI class, but it should probably be moved to a standalone environment slot, which would make the preference reloading code independent of the UI. This refactored slot could then emit a signal with the new PreferencePropertySet object as an argument; in this way, it's easy for any code that's interested in preference changes to get notified.