move camera grabber code to PreferencesPropertySet
Move the code that calls into ves::xplorer::scenegraph::SceneManager to get the current camera state out of the PreferencesTab class and into PreferencesPropertySet. This allows the names of properties to change without having to modify PreferencesTab.

This commit also includes some tweaks to the build to pull in some required header paths.