relocate near-far ratio & dragger scaling signals
Since the PreferencesPropertySet object now gets destroyed and re-created when a new .ves file is loaded, it makes sense for the PreferencePropertySet to not "own" any signals apart from those associated with live properties (so that we don't interfere with downstream slot connections when the object is destroyed).

Move the creation/registration of the 'NearFarRatio' and 'DraggerScaling' into the EventFactory, and have the PreferencePropertySet hold weak pointers to these signals, so that it may trigger them.