(WIP) add utility classes for implementing the State pattern in Squirrel
Add abstract base classes in C++ to aid in constructing state machines in Squirrel
using the State design pattern (https://en.wikipedia.org/wiki/State_pattern).

NOTE: This code doesn't actually work quite right yet. The issue is that while you
can bind a C++ class with Sqrat and extend it in Squirrel, it is not possible to
call into the derived Squirrel code given a base-class C++ pointer. Sqrat does not
automagically make this work; extra code is required on the C++ side to dynamically
look to up the correct Squirrel method and call it.

This post on the Squirrel forum offers some solutions to the problem:

http://squirrel-lang.org/forums/default.aspx?g=posts&m=7099