remove unnecessary library linkage in libves_conductor_qtui
Linking libves_conductor_qtui to Squirrel and the Squirrel standard library
is only necessary on Windows. On other platforms, the Squirrel VM code and
the standard library are statically linked into switchwire.

This small build system bug was resulting in a "copy" of the Squirrel object
code to be present in both switchwire and libves_conductor_qtui.