add Logitech Wireless Gamepad F710 support
The Logitech F710 is supported by the `xpad` driver on Linux, and is
reported as a "Generic X-Box pad" by SDL2. Make sure that the correct
X-Box 360-specific VR Juggler configuration file gets loaded for this
gamepad, since the button and analog stick layout is compatible with
the official X-Box 360 gamepad.

Hopefully this will ensure that any X-Box 360 gamepad clone will work
correctly with VE-Suite.