Add a "glue" signal to Wand to tie it into the new ProcessSelection method
Rather than changing the signatures of one or more of the button signals in
the Wand code (and risk breaking something elsewhere), as a workaround add
a "glue" signal called "Wand.SelectionButtonRelease" that is fired  when
the Wand's button 0 is released. This signal is wired into the new
Selection::ProcessSelection() slot.