(Temporarily) disable the pink glow shader effect on selected objects
When the pink and red glow shader effects overlap, it can make the red
glow hard to see, which hurts the usability of the part manipulation
feature.

A long-term solution might involve somehow figuring out which device
sent a "button released" signal, and conditionally enabling/disabling
the pink glow effect based on that (for example, the pink glow might
be enabled for a mouse, but not for a wand or gamepad).