Project the forward vector onto the ground plane
Rather than user the forward vector we get from the HMD directly,
project it onto the ground plane, so that navigation can be relative
to the ground plane rather than the HMD wearer's head. This also
allows using a "constant" up vector that is orthogonal to the ground
plane, which should allow nav rotations to behave in a consistent way,
and hopefully make VR navigation reasonably intuitive.