Handle pressing and holding the up/down nav buttons
OSVR gives us "edge-triggered" button events -- our button callback
is only invoked when the button value changes. So, instead of calling
setButton() on the MxGamePad object directly in the callback, set a
boolean "latch" variable, and move the setButton() call into
App::latePreFrame(), and make it conditional based on the value of
the boolean latch.