Remove unused C++ state machine code
This stuff breaks the build with GCC 4.8.5 on CentOS 7, so it needs to
go. I didn't end up using this code anyway -- it was written based on a
misunderstanding of how the C++ <--> Squirrel "bridge" worked, and I
eventually wrote all the state machine stuff in pure Squirrel for the
part manipulator feature branch.