More Qt uic build tweaks
The `QT4_WRAP_UI()` macro doesn't take intermediate directories into
account, so the compiled files for the ui files in the `plugin/` directory
get placed in the wrong location. We also can't control the location of
the output files (the macro hard-codes an output path), so as a workaround
just use a custom command that calls `uic` in the same way that the macro
does.