Fix shared library constructors
Both the conductor and xplorer libraries have a constructor/entrypoint
function called `vrkitLibraryInit`. On a CentOS/RHEL 7 system with a GCC
4.8.5 C++ toolchain, this causes problems -- when VE-Suite starts, the
initializer code from the conductor library runs twice, and xplorer's
initializer code does not run at all. This in turn will cause VE-Suite
to crash when loading a project (it attempts to call a method on a null
`PluginBasePtr`), because the plugin path setup stuff never runs and
VE-Suite cannot load the required plugin(s).

To address this, fix up the constructor function signatures and give
them unique names.